Refactor: Split client logic outside of app front-end
